refactor(integration): move Next.js middleware tests to dedicated file
Move percent-encoded URL, dot segment, and double slash bypass tests
from dynamic-keys.test.ts into integration/tests/nextjs/middleware.test.ts
to match the Astro and Nuxt test structure.
